+ Improved grid_copy() functions.

+++ sugar/cho_array.h
@@ -44,7 +44,7 @@ namespace cho
 		inline constexpr								array_pod									()																			noexcept	= default;
 		inline											array_pod									(uint32_t initialSize)																	{ resize(initialSize);		}
 		inline											array_pod									(::std::initializer_list<_tPOD> init)													{ 
-			throw_if(errored(resize(init.size())), ::std::exception(), "Failed to resize array! Why?");
+			throw_if(errored(resize((uint32_t)init.size())), ::std::exception(), "Failed to resize array! Why?");
 			memcpy(Data, init.begin(), Count * sizeof(_tPOD));
 		}
 		inline											array_pod									(array_pod<_tPOD>&& other)													noexcept	{
